requirements.txt
asgiref==3.8.1
attrs==23.2.0
Automat==22.10.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
constantly==23.10.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.5
cssselect==1.2.0
dj-database-url==2.1.0
Django==4.2.11
django-ckeditor==6.7.1
django-ckeditor-5==0.2.12
django-extensions==3.2.3
django-js-asset==2.2.0
django-rest-swagger==2.2.0
django-simple-settings==0.3.1
django-split-settings==1.3.1
djangorestframework==3.15.1
drf-yasg==1.21.7
filelock==3.13.4
flake8==7.0.0
greenlet==3.0.3
gunicorn==21.2.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
inflection==0.5.1
itemadapter==0.8.0
itemloaders==1.1.0
itypes==1.2.0
Jinja2==3.1.3
jmespath==1.0.1
lxml==5.2.1
MarkupSafe==2.1.5
mccabe==0.7.0
openapi-codec==1.3.2
packaging==24.0
parsel==1.9.1
pillow==10.3.0
playwright==1.43.0
Protego==0.3.1
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pycparser==2.22
PyDispatcher==2.0.7
pyee==11.1.0
pyflakes==3.2.0
pyOpenSSL==24.1.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
queuelib==1.6.2
requests==2.31.0
requests-file==2.0.0
Scrapy==2.11.1
scrapy-playwright==0.0.34
service-identity==24.1.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
tldextract==5.1.2
Twisted==24.3.0
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
w3lib==2.1.2
zope.interface==6.2
django-health-check==3.18.
celery==5.3.6
django-celery-beat==2.6.0
django-celery-results== 2.5.1
flower==2.0.1
redis==5.0.1
mysqlclient==2.2.4
psutil==5.9.8
memcachedcache==0.0.2
pymemcache==4.0.0
kombu==5.3.7
selenium==4.20.0 
PyVirtualDisplay==3.0 
beautifulsoup4==4.12.3
webdriver-manager==4.0.1
Pillow==10.3.0
webp==0.3.0 